+++ components/src/main/java/org/patternfly/component/skeleton/Skeleton.java
@@ -22,6 +22,7 @@ import org.patternfly.style.Size;
 import elemental2.dom.HTMLElement;
 
 import static org.jboss.elemento.Elements.div;
+import static org.patternfly.core.Ouia.ouiaSafe;
 import static org.jboss.elemento.Elements.span;
 import static org.patternfly.style.Classes.component;
 import static org.patternfly.style.Classes.modifier;
@@ -50,6 +51,7 @@ public class Skeleton extends BaseComponent<HTMLElement, Skeleton> {
 
     Skeleton() {
         super(ComponentType.Skeleton, div().css(component(skeleton)).element());
+        ouiaSafe(element(), false);
     }
 
     // ------------------------------------------------------ builder